requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
alembic
pydantic
pytest
requests